[fix] syntax

+++ 0x1A-hash_tables/3-hash_table_set.c
@@ -10,7 +10,7 @@
 int hash_table_set(hash_table_t *ht, const char *key, const char *value)
 {
 	hash_node_t *new, *tmp;
-	unsigned int idx;
+	unsigned long int idx;
 
 	if (ht == NULL || key == NULL || !*key)
 		return (0);
